remove duplicated Alpha/LEM benchmarks and examples, making LEM the norm
