+++ .github/workflows/bench-pr-comment.yml
@@ -35,7 +35,7 @@ jobs:
       - uses: boa-dev/criterion-compare-action@v3
         with:
           # Optional. Compare only this benchmark target
-          benchName: "fibonacci_lem"
+          benchName: "fibonacci"
           # Needed. The name of the branch to compare with
           branchName: ${{ github.ref_name }}
 
@@ -83,7 +83,7 @@ jobs:
       - uses: boa-dev/criterion-compare-action@v3
         with:
           # Optional. Compare only this benchmark target
-          benchName: "fibonacci_lem"
+          benchName: "fibonacci"
           # Optional. Features activated in the benchmark
           features: "cuda"
           # Needed. The name of the branch to compare with
